pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python-simple-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false
packages = [{include = "*", from="src"}]

[tool.poetry.scripts]
dev = "src.console:dev"
test = "tests.conftest:run"

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "2.0.29"
pydantic = {extras = ["email"], version = "^2.6.4"}
psycopg2 = "2.9.5"
asyncpg = "0.29.0"
alembic = "1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
starlette = "0.36.3"
httpx = "0.27.0"
greenlet = "3.0.3"
bcrypt = "4.0.1"
passlib = "1.7.4"
python-multipart = "0.0.9"
redis = "5.0.3"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "8.1.1"
pytest-asyncio = "0.23.6"
black = "24.3.0"
autoflake = "^2.3.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
flake8-quotes = "^3.4.0"
flake8-annotations = "^3.0.1"
flake8-comprehensions = "^3.14.0"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.1"
mccabe = "^0.7.0"
pep8-naming = "^0.13.3"
pycodestyle = "^2.11.1"
pyflakes = "^3.2.0"

[tool.black]
line-length = 120


[tool.flake8]
exclude = [".git", ".venv", "__pycache__", "env", "venv"]
inline-quotes = "double"
max-line-length = 120
ignore = "ANN101,ANN102,D200,D100,D101,D102,D107,D103,I001,I005,I004,I003"

[tool.isort]
profile = "black"
line_length = 120


[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
asyncio_mode = "auto"
